+++ src/main.py
@@ -1,13 +1,13 @@
 import logging
-from .agents.research_orchestrator import ResearchOrchestrator
+from .services.cached_research_service import CachedResearchService
 from config import GEMINI_API_KEY
 
 request_to_agent: str = "Nvidia stock price"
 def main():
     if not GEMINI_API_KEY:
         raise Exception()
-    orchestrator = ResearchOrchestrator(GEMINI_API_KEY)
-    research_results: str = orchestrator.run(request_to_agent)
+    service = CachedResearchService(GEMINI_API_KEY)
+    research_results: str = service.run(request_to_agent)
     print(research_results)
 
 if __name__ == "__main__":